requirements.txt
asgiref==3.11.1
Django==5.2.11
django-formtools==2.5.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
djangorestframework==3.15.2